Fix two bugs
* Fix keyboard covered text field when in focus
* Fix database deletion restart
* Refactor config screens

Related to:
OPTM-60, OPTM-81
