+++ lib/screens/setup/welcome_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:p2p_task/screens/setup/config_view.dart';
+import 'package:p2p_task/screens/setup/config_screen.dart';
 import 'package:p2p_task/screens/setup/sync_config_screen.dart';
 import 'package:p2p_task/services/change_callback_notifier.dart';
 import 'package:p2p_task/services/identity_service.dart';
@@ -13,39 +13,37 @@ class WelcomeScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: ConfigView(
-        title: 'Welcome to the P2P Task Manager App 👋',
-        onSubmit: () => _handleSubmit(context),
-        child: Form(
-          key: formKey,
-          child: Column(
-            children: [
-              Row(
-                children: [
-                  Text(
-                    'Choose a name:',
-                  ),
-                ],
-              ),
-              TextFormField(
-                validator: (value) {
-                  if (value == null || value.isEmpty || value.length < 3) {
-                    return 'Name must be at least 3 characters long.';
-                  }
-
-                  return null;
-                },
-                decoration: InputDecoration(
-                  helperText:
-                      'This name will be shown to other users when they connect with this device.',
-                  helperMaxLines: 2,
+    return ConfigScreen(
+      title: 'Welcome to the P2P Task Manager App 👋',
+      onSubmit: () => _handleSubmit(context),
+      child: Form(
+        key: formKey,
+        child: Column(
+          children: [
+            Row(
+              children: [
+                Text(
+                  'Choose a name:',
                 ),
-                controller: nameController,
-                onFieldSubmitted: (value) => _handleSubmit(context),
+              ],
+            ),
+            TextFormField(
+              validator: (value) {
+                if (value == null || value.isEmpty || value.length < 3) {
+                  return 'Name must be at least 3 characters long.';
+                }
+
+                return null;
+              },
+              decoration: InputDecoration(
+                helperText:
+                    'This name will be shown to other users when they connect with this device.',
+                helperMaxLines: 2,
               ),
-            ],
-          ),
+              controller: nameController,
+              onFieldSubmitted: (value) => _handleSubmit(context),
+            ),
+          ],
         ),
       ),
     );
